intense refactoring.
