+++ app.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python3
+from controllers import MainController
+
+# Run the application.
+if __name__ == "__main__":
+    MainController().loop()
