+++ controllers/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3
+from .base_controller import BaseController
+from .base_actor import BaseActor
+from .human_actor import HumanActor
+from .machine_random_actor import MachineRandomActor
+from .main_controller import MainController
+name = 'controllers'
