--- definitions.py
@@ -1,13 +1,0 @@
-#!/usr/bin/python3
-
-
-class Player:
-    USER = 1
-    OPP = -1
-
-
-class FieldValue:
-    USER = 1
-    OPP = -1
-    EMPTY = 0
-    POSSIBLE_MOVE = 2
