+++ games/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/python3
+from .board_game import BoardGame
+from .bauernschach import Bauernschach
+from .alquerque import Alquerque
+name = 'games'
