+++ gui/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/python3
+from .output import Output
+name = 'gui'
