+++ gui/output.py
@@ -2,28 +2,7 @@
 
 import re
 import os
-from definitions import Player, FieldValue
-
-
-class ScreenParameters:
-    """
-    Pass parameters from game loop to screen rendering:
-    - bord: 2D array for rows and cells/cols.
-    - moveHistory: array of 2-tuples of 2-tuples
-      [((fromRow, fromCol), (toRow, toCol)), ...].
-    - currentPlayer: int constant identifying the player.
-    - feedback: feedback from executing the previous user input.
-    - question: instruction for the user.
-    - options: dictionary of possible cli-inputs and a matching description.
-    """
-
-    game = None
-    board = []
-    moveHistory = []
-    player = ""
-    feedback = ""
-    instruction = ""
-    options = {}
+from models import FieldValue, Player
 
 
 class Output:
@@ -114,8 +93,10 @@ class Output:
             "History:       ",
         ]
         for step in steps:
-            history.append(self._map_field_coordinates_to_text(
-                step[0]) + " to " + self._map_field_coordinates_to_text(step[1]))
+            fieldFrom = self._map_field_coordinates_to_text(step[0])
+            fieldTo = self._map_field_coordinates_to_text(step[1])
+            entry = fieldFrom + " to " + fieldTo
+            history.append(entry)
         return history
 
     def _prepare_player(self, player):
