--- machine.py
@@ -1,5 +1,0 @@
-#!/usr/bin/python3
-
-
-class RandomMachine:
-    pass
