+++ models/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python3
+from .current_game import CurrentGame
+from .definitions import Player, FieldValue, Games, States, Commands
+from .field import Field
+from .screen_parameters import ScreenParameters
+name = 'models'
