+++ models/current_game.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python3
+class CurrentGame:
+    # Game setup.
+    board_size = 4
+    game_choice = None
+    machine = None
+    player_to_start = None
+
+    # Game engine, state, players
+    game = None
+    current_actor = None
+    waiting_actor = None
+    history = []
